requirements.txt
# Streamlit para o front-end
streamlit>=1.24.0
streamlit-lottie>=0.0.5

# APIs de IA
openai>=1.3.0
google-generativeai>=0.1.0
requests>=2.31.0

# Manipulação de dados e exportação
pandas>=2.1.0
numpy>=1.26.0
xlsxwriter>=3.1.0
fpdf2>=2.7.0

# Leitura de PPT para playbook
python-pptx>=0.6.21

# Gerenciamento de variáveis de ambiente
python-dotenv>=1.0.0

# Tokens (necessário para ChatGPT)
tiktoken>=0.4.0
